Changed some styling.
Version: 0.0.2
Author(s): Anthony Pizzimenti
Date: July 6, 2017

========================= MAJOR CHANGES ===========================
Shifts the styling around in the html retrieved from Hacker News.
Makes sure that, before the HTML is rendered, padding is adjusted
for all rows in the table, and adds a nicely colored header (with
the same color as Hacker News).

I hopped on the bandwagon (and Facebook's, apparently) and did an
ASCII banner for when the script starts up. Includes the version.

=========================== FEATURES ==============================
Nothing of note.